requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
python-multipart
markdown
mammoth
requests
pgvector
scikit-learn
nltk
python-docx
PyPDF2
beautifulsoup4
sumy
langdetect
deep-translator
python-dotenv
google-cloud-aiplatform
